som-external: Make chrony sources configurable by WebLCM
This change creates and enables the use of a chrony .sources file
(/etc/chrony/supplemental.sources) which is utilized by WebLCM to expose the
functionality to configure NTP sources for chrony.

When used with the lrd-encrypted-storage-toolkit, the supplemental.sources file
is relocated to /data/secret - as is done with similar "runtime"-changeable
configuration files (e.g., WebLCM, NetworkManager, stunnel).

Also added here is modular support in WebLCM for configuring chrony sources.

Bug 22665