feat: Switch to dm-crypt for rodata encryption
This change switches from using fscrypt to dm-crypt when generating and mounting
the encrypted read-only data image (rodata.img). The new mounting script expects
the encryption key to be present in the kernel keyring as a logon key with a
description of "dm-crypt:ffffffffffffffff", and it uses this key to decrypt
`rodata.img` and mount it at `/rodata`.

Refs: PROD-27936