feat: Support AWS KMS signing via PKCS#11
This change introduces a new feature to support signing images via the AWS Key
Management Service (KMS). The main benefit here is that the private key material
does not need to be stored locally on the build machine and is never exposed
outside of AWS KMS.

Two new packages are introduced here: `pkcs11-provider` and `aws-kms-pkcs11`.

`pkcs11-provider` is an OpenSSL provider which enables the use of PKCS#11 APIs
for cryptographic operations.

`aws-kms-pkcs11` is a project which exposes a PKCS#11 API for interfacing with
AWS KMS.

When used together, images can be signed with a key stored in AWS KMS.

Newly introduced environment variables:
- `AWS_KMS_SIGNING`: Set to `y` to enable AWS KMS signing logic
- `AWS_KMS_SIGNING_PKCS11_PUBLIC_CERT_PATH`: The path to the local copy of the
public key which corresponds to the signing public key. This is also assumed to
be the "keys" directory.
- `AWS_KMS_PKCS11_SLOT_LABEL`: Target AWS KMS key alias (or label)
- `AWS_KMS_PKCS11_KMS_KEY_ID`: Target AWS KMS key ID
- `AWS_KMS_PKCS11_AWS_REGION`: Target AWS KMS key AWS region

Refs: PROD-45023